All the core media calculation stuff was hosing us.  This works perfectly.  Need to refactor
