+++ Sources/Buffie/File Writers/HLS/Atoms/TRAF.swift
@@ -8,21 +8,19 @@ struct TRAF: BinarySizedEncodable {
     var trackDecodeAtom: [TFDT] = [TFDT()]
     var trackRun: [TRUN] = [TRUN()]
     
-    static func from(_ samples: [Sample], previousDecodeTime: UInt64) -> [TRAF] {
+    static func from(_ samples: [Sample], previousDuration: UInt64) -> [TRAF] {
         
         var traf = TRAF()
         if let sample = samples.first {
             traf.trackFragmentHeader = [TFHD.from(sample: sample)]
+            
+            let duration = samples.reduce(0) { (cnt, sample) in cnt + sample.duration.value }
+            traf.trackDecodeAtom = [TFDT.from(decode: UInt64(sample.decode.value),
+                                              duration: UInt64(duration) )]
         }
-        
-        let duration = samples.reduce(kCMTimeZero) { (cnt, sample) in
-            CMTimeAdd(cnt, sample.duration)
-        }
-        
-        traf.trackDecodeAtom = [TFDT.from(decode: UInt64(samples.first!.decode.value),
-                                          duration: UInt64(duration.value) + previousDecodeTime )]
         
         traf.trackRun = [TRUN.from(samples: samples)]
+        
         return [traf]
     }
