+++ Sources/Buffie/File Writers/HLS/HLSPlaylist.swift
@@ -1,4 +1,5 @@
 import Foundation
+import CoreMedia
 
 enum HLSPlaylistType: String {
     case vod   = "VOD"
@@ -71,7 +72,7 @@ class HLSPlaylistWriter {
     
     func write(segment: FragmentedMP4Segment) {
         if let name = segment.file.path.components(separatedBy: "/").last {
-            self.writeSegment(named: name, with: segment.duration)
+            self.writeSegment(named: name, with: CMTimeGetSeconds(segment.duration))
         }
     }
