+++ Sources/Buffie/File Writers/HLS/Sample.swift
@@ -1,0 +1,39 @@
+import Foundation
+import CoreMedia
+
+public struct Sample {
+    
+    var type: SampleType
+    var format: CMFormatDescription
+    var nalus: [NALU] = []
+    var duration: CMTime
+    var pts: CMTime
+    var decode: CMTime
+    
+    var size: UInt32 {
+        return self.nalus.reduce(0, { last, nalu in last + nalu.totalSize })
+    }
+    
+    var dependsOnOthers: Bool = false
+    var isSync: Bool = false
+    var earlierDisplayTimesAllowed: Bool = false
+    
+    init(sampleBuffer: CMSampleBuffer) {
+        self.type       = .video
+        self.format     = CMSampleBufferGetFormatDescription(sampleBuffer)!
+        self.duration   = CMSampleBufferGetDuration(sampleBuffer)
+        self.pts        = CMSampleBufferGetPresentationTimeStamp(sampleBuffer)
+        self.decode     = CMSampleBufferGetDecodeTimeStamp(sampleBuffer)
+        
+        self.isSync                     = !sampleBuffer.notSync
+        self.dependsOnOthers            = sampleBuffer.dependsOnOthers
+        self.earlierDisplayTimesAllowed = sampleBuffer.earlierPTS
+        
+        if let bytes = bytes(from: sampleBuffer) {
+            for nalu in NALUStreamIterator(streamBytes: bytes, currentIdx: 0) {
+                self.nalus.append(nalu)
+            }
+        }
+    }
+    
+}
