+++ Tests/BuffieTests/FragmentedMP4WriterTests.swift
@@ -57,17 +57,17 @@ class FragmentedMP4WriterTests: XCTestCase {
         XCTAssertEqual(Int(results.first!.payloadSize), compressedFrame.count-4)
     }
     
-//    func test_experiment() {
-//        let dir    = URL(fileURLWithPath: "/tmp")
-//        let writer = try? FragmentedMP4Writer(dir)
-//        XCTAssertNotNil(writer)
-//
-//        let reader = SimpleReader() { sample in writer?.got(sample) }
-//        let camera = try? Camera(.back, reader: reader, controlDelegate: nil)
-//        XCTAssertNotNil(camera)
-//        camera?.start()
-//        let e = self.expectation(description: "Blah")
-//        self.wait(for: [e], timeout: 60)
-//    }
+    func test_experiment() {
+        let dir    = URL(fileURLWithPath: "/tmp")
+        let writer = try? FragmentedMP4Writer(dir)
+        XCTAssertNotNil(writer)
+
+        let reader = SimpleReader() { sample in writer?.got(sample) }
+        let camera = try? Camera(.back, reader: reader, controlDelegate: nil)
+        XCTAssertNotNil(camera)
+        camera?.start()
+        let e = self.expectation(description: "Blah")
+        self.wait(for: [e], timeout: 20)
+    }
     
 }
